Added the flight scheduler
