+++ src/main/java/virtualdispatcher/VirtualDispatcherApplication.java
@@ -1,35 +1,36 @@
 package virtualdispatcher;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import virtualdispatcher.core.scheduling.FlightScheduler;
+
+import java.util.Timer;
+import java.util.TimerTask;
 
 @SpringBootApplication
-public class VirtualDispatcherApplication {
+public class VirtualDispatcherApplication implements ApplicationRunner {
+    @Autowired
+    private FlightScheduler flightScheduler;
 
     public static void main(final String[] args) throws Exception {
         SpringApplication.run(VirtualDispatcherApplication.class, args);
     }
 
-    //@Override
-    public String getName() {
-        return "VirtualDispatcher";
-    }
+    @Override
+    public void run(ApplicationArguments args) throws Exception {
+        TimerTask schedulerTask = new TimerTask() {
+            @Override
+            public void run() {
+                flightScheduler.scheduleFlights();
+            }
+        };
 
-//    @Override
-//    public void initialize(final Bootstrap<VirtualDispatcherConfiguration> bootstrap) {
-//        bootstrap.addBundle(new AssetsBundle("/assets", "/", "flightStatus.html"));
-//    }
-//
-//    @Override
-//    public void run(final VirtualDispatcherConfiguration configuration,
-//                    final Environment environment) {
-//        // Serve API resources at /api path
-//        environment.jersey().setUrlPattern("/api/*");
-//
-//        // Run the application
-//        Guice
-//            .createInjector(new ApplicationModule(environment, configuration))
-//            .getInstance(ApplicationRunner.class)
-//            .run(environment, configuration);
-//    }
+        Timer timer = new Timer("Flight Scheduler Timer");
+
+        long intervalMilli = 3000L;
+        timer.scheduleAtFixedRate(schedulerTask, intervalMilli, intervalMilli);
+    }
 }
